reset stats on POST /_restart and correct open_databases test.  Closes COUCHDB-398
I rearranged the shutdown order from what was committed in the 0.9.x branch so
that the stats collector is still up when couch_server shuts down.  This way the
messages to decrement the number of open_databases during the shutdown don't
fail with a badarg. The badarg is silently ignored in 0.9.x, but here we pattern
match with a true so we end up with lots of error reports.

You may see some spurious erlang error reports when running the test suite on
an SMP emulator, due to the use of public ets tables in the stats collector API.


git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@789829 13f79535-47bb-0310-9956-ffa450edef68