need to add sha1.js to allow account creation on status page
it doesn't really make sense to create accounts on a page that
will show nothing unless you are logged in under an admin account, but
so it goes.  Better would be to hide the "status" link unless you're
logged in as admin.

but anyway adding this js lib is relatively harmless, and eliminates
that ugly dialog error

modified:   share/www/status.html

git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@900567 13f79535-47bb-0310-9956-ffa450edef68