safer db close on rep terminate than in r1133561
As fdmanana pointed out to me, my last change to this broke the
replication js test. While dropping the ref on a local database is not
strictly necessary, closing a remote database might terminate an http
connection that is being help open. The remote server will complain
if we drop it unexpectedly.

git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1135313 13f79535-47bb-0310-9956-ffa450edef68