Fix server crash associated with the replicator database
If there's an exception when calculating the replication ID
for a replication document, it crashes the replication manager
gen_server 10 times. 10 is the maximum number of restarts per
hour specified for the couch_server_sup supervisor.
An easy way to trigger such exception is to specify a non
existent filter in a replication document.

This closes COUCHDB-1199.



git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1137928 13f79535-47bb-0310-9956-ffa450edef68