Make replicator agnostic about the update seq type
To determine which row, coming from a _changes feed, came
earlier (or later), the replicator compared their respective
update seq field which is a number in Apache CouchDB. This
worked fine, as rows coming from the _changes stream are
sorted by they're update seq field. However it didn't work
when the update seq is not a number, such as in the case of
BigCouch for example (strings).

This change allows for pull replicating from a BigCouch cluster,
simplifies some code and makes it more robust overall.



git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1144419 13f79535-47bb-0310-9956-ffa450edef68