more portability for build search paths
With this change, the default search paths for headers and libraries
are appended so that they do not override any environment variables
either exported or set on the command line, as per ./configure --help.

Also, don't add these paths for windows builds. It can confuse the
selection of the proper ssl libs.

git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1146999 13f79535-47bb-0310-9956-ffa450edef68