Allow "versioned" names of the auto-tools to be used.
Previously bootstrap required aclocal/automake to be available as
"aclocal" and "automake", but some platforms allows multiple versions
of these tools to be installed and makes them available as
"aclocal-<version>".

This patch extends the bootstrap script to try to search for various
names for the tools.

Patch by Trond Norbye.

git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1155352 13f79535-47bb-0310-9956-ffa450edef68