Add configurable automatic compaction
A set of rules can now be defined in order to automatically
trigger the compaction of databases and their views. This
configuration can be global or overrided for specific
databases. By default it's disabled. Enabling it can be
done via the .ini configuration by simply adding global
or database specific compaction rules.

Closes COUCHDB-1153.
Thanks everyone involved who gave feedback and suggestions
for improvements.



git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1159922 13f79535-47bb-0310-9956-ffa450edef68