Revert "store the port in the cached config. Useful to get it when it was set"
When setting the ssl port, the server was crashing. This is due to the
nature config changes are handled (when a config change happen, the
process is exited). At some point we need to fully rewrite couch_config
to only reload a server configuration when an event is coming rather
than crashing it.

This reverts commit 174025eb0e069b86f0b980f1b27223b49464f109.