`_view` filter is now hanled by couch_mrview:view_changes_since/7
Rather than passing all doc sent y couch_db:changes_since/5 to a
transformed view function we are now using the more efficient
`couch_mrview:view_changes_since/7` wich use the seq btree in the view
group.

Also add the parameters `view_filter=DName/FilterName` which allows to
filter any view changes using the fiter function `FilterName` from the
design document `DName`:

   GET http://<node>/<dbname>/_changes?filter=_view&view=<viewname>&view_filter<filtername>

Where variables are:

- <node>: URL of the CouchDB node
- <dbname>: Name of the database
- <viewname>: DesignId/ViewNameThe name of the view in the views properties
  from the design document DesignId`
- <filtername>: DesignId1/FilterName The name of the filter in
  the filters properties from the design document DesignId1

The difference with the old behaviour is that you really accessing the
view index instead of passing each map function to all the changed docs
wich is a way more efficient (you do'nt open a system process, serialize
the function, pass it to the process, wait for return, test it and
eventually return a change). You can also filter the results. Other
advantages are that the views are refreshed in real time when the
database content changes and the view changes are only triggered if the
index is updated.