Fix random failures in replication.js test
This should fix the random failures for the 'expected "25" got "31"
error message that pops up occasionally. If my hunch is correct when we
remove the two docs from the source db its possible to get the writes on
the target out of order depending on a race condition.

This just sleeps a bit in between the deletions to try and reduce that
possibility. I haven't spotted this error after applying this patch.