add `seq_indexed` option to a design document.
If the option `seq_indexed` is **true**, the views indexes will be also
indexed by sequences. It is possible to retrieve changes using the
`couch_mrview:view_changes_since/{5,6,7}`  function wich allows you to
get all changes in a view or changes from a view in a range since last
sequence. ex:

    couch_mrview:view_changes_since(Db, <<"_design/test">>, <<"test">>,
0, fun(KV, Acc) -> io:format("kv ~p~n", [KV]), {ok, Acc} end,
[{start_key, null}, {end_key, null}]).
    kv {{null,4},
        {<<"1f7d72c4ae8cab487c7efeff050011fe">>,
         {[{<<"_id">>,<<"1f7d72c4ae8cab487c7efeff050011fe">>},
           {<<"_rev">>,<<"1-9b2a5b981a9dd0ad1b3a3e826498e9bb">>},
           {<<"test">>,null}]}}}
    {ok,[]}

If the range is omitted all changes will be retrieved since the last
sequence.

Internals:

When `seq_indexed` is created 1 main sequence is created for all the
view group with the key `{ViewId, Seq}` and the value `{DocId, Key,
Val}`
is added. There can be multiple lines by sequences. These sequences are
retrieved by querying the b-tree in the range:  `[{ViewId, StartSeq+1},
{ViewId, EndSeq}]`
where EndSeq depending on the direction can be 0 or a the last big
integer.

To allows the querying of changes in a range, a new index is added per
views with the composite key `{Key, Seq}` (instead DocId) and the value
`{DocId, Val}` is indexed for this key. We query this index in the range
`[{StartKey, StartSeq+1}, {EndKey, EndSeq}]` to retrieve the changes.

Caveats:

The size of the indexes will grow significantly with this changes and
compaction is needed more often.