handle couch_mrview events.
This change introduces an efficient way to manage couch_mrview events
that notify a suscribed process about an index update.

Internals:

When a process subscribe to a an index updated using
`couchdb_mrview_events:subscribe/3` it open a new indexer that will
refresh the indexes each time the database is updated. When no more
subscribers exists the automatic indexer is killed. Each indexer is
supervised.

The subscribe receive 2 kinds of notifications:

- `{index_update, {DbName, DDocName}}` when an index is updated
- `{index_shutdown, {DbName, DDocName}}` when the db or the index are
deleted.

Ex:

    1> couch_mrview_events:subscribe(<<"testdb">>, <<"_design/test">>).
    ok
    2> ok
    2>

    21:11:44.673 [info] 127.0.0.1 - - POST /testdb 201
    21:11:44.698 [info] Opening index for db: testdb idx: _design/test sig: "8152471699732c18e2a4954ac27b1049"
    21:11:44.699 [info] Starting index update for db: testdb idx: _design/test
    21:11:44.737 [info] Index update finished for db: testdb idx: _design/test

    2> flush().
    Shell got {index_update,{<<"testdb">>,<<"_design/test">>}}
    ok
    3>

    21:12:09.799 [info] 127.0.0.1 - - DELETE /testdb/_design/test?rev=1-114ec5224d4dc5d63912cd50536bdf48 200
    21:12:09.801 [info] Closing index for db: testdb idx: _design/test sig: "8152471699732c18e2a4954ac27b1049"

    3> flush().
    Shell got {index_shutdown,{<<"testdb">>,<<"_design/test">>}}
    ok