Signficantly simplify the auth cache
The old cache was overly complicated for doing what it does. This new
version does introduce a race condition on cache invalidation because
the update has to go through the even notification and then the changes
feed needs to be processed. In the future we should generalize the db
hook functions and turn it into a write through cache. Until then I just
added a short wait in the tests.

patch from @davisp edited for rcouch.