workaround to prevent the view changes reader to crage on non found doc
when a doc is removed from the database in the current version of the
view changes index, this change is not indexed and the view change index
has still the key indexed. The way to delete for now is to use
include_deleted=true and push deleted doc that match the view so the
deleted document can be indexed as deleted.

But when the doc is purged and someone fetch the changes the view
changes reader was crashing because it didn't find the doc. This
workaround patch the changes reader to ignore the documents not found
and simply log them.