Don't die when a known replicator related pid dies
The couch_replicator_manager would die when any replication related pid
exited. Instead of that we don't. This means we don't cancel and restart
all known replications that happen to be running on the node due to an
error starting replications for a given database or document.

fix from couchdb 1.6.1
https://github.com/apache/couchdb-couch-replicator/commit/f90b6e0ee0dceca7d43c1694353e2784afb92cb6